pyproject.toml
[project]
name = "softdesk"
version = "1.0.0"
description = "Application SoftDesk (Django + DRF)"
authors = [{ name = "r3n3gat" }]
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
  "django (>=5.0,<6.0)",
  "djangorestframework (>=3.15,<4.0)",
  "djangorestframework-simplejwt (>=5.3,<6.0)",
  "python-decouple (>=3.8,<4.0)",
  "drf-yasg (>=1.21,<2.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
